fix(server): parse operator monitor workflow config once
